Fix issues on 32-bit systems introduced by r88460

+++ Modules/zlibmodule.c
@@ -951,10 +951,10 @@ PyZlib_adler32(PyObject *self, PyObject *args)
         Py_BEGIN_ALLOW_THREADS
         /* Avoid truncation of length for very large buffers. adler32() takes
            length as an unsigned int, which may be narrower than Py_ssize_t. */
-        while (len > (Py_ssize_t)UINT_MAX) {
+        while (len > (size_t) UINT_MAX) {
             adler32val = adler32(adler32val, buf, UINT_MAX);
-            buf += UINT_MAX;
-            len -= UINT_MAX;
+            buf += (size_t) UINT_MAX;
+            len -= (size_t) UINT_MAX;
         }
         adler32val = adler32(adler32val, buf, len);
         Py_END_ALLOW_THREADS
@@ -989,10 +989,10 @@ PyZlib_crc32(PyObject *self, PyObject *args)
         Py_BEGIN_ALLOW_THREADS
         /* Avoid truncation of length for very large buffers. crc32() takes
            length as an unsigned int, which may be narrower than Py_ssize_t. */
-        while (len > (Py_ssize_t)UINT_MAX) {
+        while (len > (size_t) UINT_MAX) {
             crc32val = crc32(crc32val, buf, UINT_MAX);
-            buf += UINT_MAX;
-            len -= UINT_MAX;
+            buf += (size_t) UINT_MAX;
+            len -= (size_t) UINT_MAX;
         }
         signed_val = crc32(crc32val, buf, len);
         Py_END_ALLOW_THREADS
